backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
joinly-client
fastmcp
langchain
langgraph
langchain-openai
langchain-mcp-adapters
spacy
pdfminer.six
python-docx
nltk
openai<2.0.0
python-dotenv
langchain-google-genai
psycopg2-binary
python-gitlab
langchain-community
ddgs
